+++ Ss6PlayerExamples/Plugins/SpriteStudio6/Source/SpriteStudio6/Private/SSSDK/ssplayer_cellmap.h
@@ -109,7 +109,7 @@ public:
 			delete itr.Value();
 		}
 
-		for ( size_t i = 0 ; i < CellMapList.Num(); i++ )
+		for ( int i = 0 ; i < CellMapList.Num(); i++ )
 		{
 			delete CellMapList[i];
 		}
